Changes to GIF support:
- the Nescape loop extension block is written if gif_loop is set.
This requires libungif/libgif 4.1.3 or later.

Resolves: http://rt.cpan.org/Ticket/Display.html?id=21185

- we now set the GIF header depending on the features enabled in the
GIF.  Since we don't do configure type probes for bugs we have to
enable this based on the gif_lib.h version number, it's enabled for
4.1.0 and later, but requires 4.1.3 and later (maybe 4.1.3) to run
without crashing.

You can avoid this crash by passing --nogifsetversion to Makefile.PL,
or by installing a non-buggy libungif/libgif, if there is such a
thing.