[rt.cpan.org #64785] GIF encoding is very slower on Imager 0.78+
The various backward compatibility hacks in Imager.pm's gif handling
were removed with the switch to a separate module, but this means that
the original default of using a median cut palette builder was removed.

The addi algorithm is slow enough that I'd prefer that people have to
select it manually, so change the global quant default to median cut.