avoid ignoring the result of i_io_getc()
Coverity complained about *(ig->read_ptr++) because the result of the *
operator was ignored, which was deliberate.

I could simply had added a void cast, but given i_io_peekc(),
I could see a macro that skips the * being simpler to read, adding
i_io_nextc(), which is exposed as the nextc() method on Imager::IO
objects.