re-work the mingw runner
This previously installed strawberry perl itself and updated the path
using the old add-path mechanism, which has been removed.

strawberry perl is now installed by default, which saves some time,
but the GITHUB_PATH mechanism doesn't appear to add a directory to the
front of PATH when that directory already exists in PATH.

This meant the build was using some random gcc earlier in the path,
causing build failures.

So, update the path manually for each step that needs it.