Imager now requires a 16-bit integer type
This is used for 16-bit images, previously this would fallback to
byte access, but this could cause byte order problems for
i_img_data().

The code still uses the access macros, so if some non-obsolete
platform doesn't have a 16-bit unsigned integer I can re-introduce
the byte access, but with proper byte ordering, but this seems
unlikely.