special case fetching a single linear channel
This eliminates an inner loop for this case, I was hoping it would
be autovectorized with vgather* variants, but the optimizer only
seems to handle fixed steps.

It's useful anyway, and hopefully if other code is modified to
support autovectorization (fetching all samples for a channel to
support that) this change may support that.