i_convert(): add a linear sample implementation and extra channels support
This retains the gamma oriented implementation, but the default is
linear.

This also tries to be structured to support compiler auto-vectorization
but you may need to increase the optimization level and set an
appropriate -march value.

This also adds _assert suffix versions of i_gslin(), i_gslinf()
functions, and makes these and the existing i_gsampf?_assert()
functions use assert() instead of im_fatal(), so they only apply
when assertions are enabled.