TIFF: tag types clean up
Unfortunately the libtiff interface for setting/getting tags isn't
type safe, so I went through and checked the types supplied for each
tag.

I think the types supplied for the image width/height were incorrect
for 64-bit platforms, but due to the ABI I don't think it made a
difference on amd64 (the values are passed in registers).

It may have beed a problem on other ABIs, especially big-endian
systems.