400-basic: only test the formats we configure this run
in #536 one problem was libtiff failed to be configured, but the test
picked up the already installed Imager::File::TIFF, but that version
of Imager::File::TIFF was buggy WRT to a libtiff security update,
which has presumably been installed.

So only test the current set of modules.

Part of a fix for #536